requirements.txt
Flask==2.2.3
python-dotenv==1.0.0
supabase==1.0.3
gunicorn==20.1.0
Werkzeug==2.2.3 